Show sources under expander
